safari-600.7-branch/LayoutTests/http/tests/xmlhttprequest/xmlhttprequest-overridemimetype-content-type-header-expected.txt

This tests that XMLHttpRequest overrideMimeType() properly updates the Content-Type header for the response.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS xhr.getResponseHeader("Content-Type") is "text/xml;charset=GBK"
PASS successfullyParsed is true

TEST COMPLETE
